requirements-ilpycon.txt
django-heroku
django-model-utils
django-sitetree
django-taggit
django-timezone-field
django-timezones
easy-thumbnails
gunicorn
html5lib
markdown
Pillow
pinax-boxes==3.0.1
pinax-pages==0.5.0
pytz
pinax-blog==7.0.3
pinax-images==3.0.1
